Refatorar barra de busca de médicos
